Added Prints of Lists
Added methods to print lists of to be added , to be deleted ,  added and deleted on the console and to a txt file. Also created a method to print on the console the events for every organizer but didn't finish the method to be able  to also print to the txt file.